update: move cmd-tools
