+++ cmd-tools/cfg.ts
