+++ cmd-tools/lib/bootstrap.ts
