+++ cmd-tools/lib/burn.ts
