+++ cmd-tools/lib/utils.ts
