+++ package.json
@@ -27,13 +27,11 @@
     "serve-debug": "nodemon --inspect dist/server.js",
     "watch-debug": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run serve-debug\"",
     "swap": "./swap.sh",
-    "swap:dev": "./swap.sh dev config/.default.json",
-    "proof": "ts-node ./scripts/relay.ts"
+    "swap:dev": "./swap.sh dev config/.default.json"
   },
   "dependencies": {
     "@polkadot/api": "^1.8.0-beta.2",
     "@polkadot/util": "^2.6.2",
-    "@types/chalk": "^2.2.0",
     "async": "^3.1.0",
     "bcrypt-nodejs": "^0.0.3",
     "bluebird": "^3.5.5",
@@ -94,18 +92,13 @@
     "@typescript-eslint/eslint-plugin": "^2.3.1",
     "@typescript-eslint/parser": "^2.3.1",
     "chai": "^4.2.0",
-    "chalk": "^3.0.0",
     "concurrently": "^4.1.2",
     "eslint": "^6.4.0",
     "husky": "^4.2.3",
     "jest": "^24.9.0",
-    "knex": "^0.20.13",
     "node-sass": "^4.12.0",
     "nodemon": "^1.19.2",
-    "prompts": "^2.3.2",
-    "readline": "^1.3.0",
     "shelljs": "^0.8.3",
-    "sqlite3": "^4.1.1",
     "supertest": "^4.0.2",
     "ts-jest": "^24.1.0",
     "ts-node": "^8.4.1",
